backend/requirements.txt
fastapi==0.109.0
uvicorn==0.26.0
mangum==0.17.0
openai==1.10.0
pydantic==2.5.3
python-dotenv==1.0.1
azure-search-documents==11.4.0
aiohttp==3.9.1
